Add ISSN Support
This commit adds support for ISSN identifier lookup in CORAL. The plugin
will now iterate over the combined list of ISBNs and ISSNs from Koha until
a match is found in CORAL (if there is one). In the event of multiple
possible match points between Koha and CORAL records, the first to appear in
the MARC record will be the one used.